Have 10up toolkit generate the php file to include fast refresh
