+++ packages/toolkit/config/fast-refresh.php
@@ -1,0 +1,60 @@
+<?php
+/**
+ * This file is generated by 10up-toolkit and copied into your dist folder, do not modifies this file directly.
+ */
+namespace TenUpToolkit;
+
+if ( defined( 'SCRIPT_DEBUG') && SCRIPT_DEBUG ) {
+	add_action( 'init', __NAMESPACE__ . '\\register_react_fast_refresh', 1 );
+}
+
+/**
+ * Register React Fast Refresh scripts
+ *
+ * @return void
+ */
+function register_react_fast_refresh() {
+	if ( ! defined( 'TENUP_TOOLKIT_DIST_URL' ) || ! defined( 'TENUP_TOOLKIT_DIST_PATH' ) ) {
+		trigger_error( '10up-toolkit error: you must defined both TENUP_TOOLKIT_DIST_URL and TENUP_TOOLKIT_DIST_PATH to get fast refresh to work' );
+		return;
+	}
+
+	$react_fast_refresh_entry   = TENUP_TOOLKIT_DIST_URL . 'fast-refresh/react-refresh-entry/index.min.js';
+	$react_fast_refresh_runtime = TENUP_TOOLKIT_DIST_URL . 'fast-refresh/react-refresh-runtime/index.min.js';
+
+	$react_fast_refresh_entry_path   = TENUP_TOOLKIT_DIST_PATH . 'fast-refresh/react-refresh-entry/index.min.js';
+	$react_fast_refresh_runtime_path = TENUP_TOOLKIT_DIST_PATH . 'fast-refresh/react-refresh-runtime/index.min.js';
+
+
+	if ( ! file_exists( $react_fast_refresh_entry_path ) || ! file_exists( $react_fast_refresh_runtime_path ) ) {
+		return;
+	}
+
+	wp_register_script(
+		'react-fast-refresh-entry',
+		$react_fast_refresh_entry,
+		[],
+		filemtime( $react_fast_refresh_entry_path ),
+	);
+
+	wp_register_script(
+		'react-fast-refresh-runtime',
+		$react_fast_refresh_runtime,
+		[],
+		filemtime( $react_fast_refresh_runtime_path ),
+	);
+
+	global $wp_scripts;
+	$script = $wp_scripts->query( 'react', 'registered' );
+
+	if ( $script ) {
+		if( !in_array( 'react-fast-refresh-runtime', $script->deps ) ){
+			$script->deps[] = 'react-fast-refresh-runtime';
+		}
+
+		if( !in_array( 'react-fast-refresh-entry', $script->deps ) ){
+			$script->deps[] = 'react-fast-refresh-entry';
+		}
+
+	}
+}
