+++ projects/10up-theme/functions.php
@@ -14,6 +14,12 @@ define( 'TENUP_THEME_DIST_URL', TENUP_THEME_TEMPLATE_URL . '/dist/' );
 define( 'TENUP_THEME_INC', TENUP_THEME_PATH . 'includes/' );
 define( 'TENUP_THEME_BLOCK_DIR', TENUP_THEME_INC . 'blocks/' );
 
+if ( file_exists( __DIR__ . '/dist/fast-refresh.php' ) ) {
+	define( 'TENUP_TOOLKIT_DIST_URL', TENUP_THEME_DIST_URL );
+	define( 'TENUP_TOOLKIT_DIST_PATH', TENUP_THEME_DIST_PATH );
+	require_once __DIR__ . '/dist/fast-refresh.php';
+}
+
 require_once TENUP_THEME_INC . 'core.php';
 require_once TENUP_THEME_INC . 'overrides.php';
 require_once TENUP_THEME_INC . 'template-tags.php';
